requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
diff-match-patch==20241021
Django==5.1.7
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-crispy-forms==2.3
django-import-export==4.3.7
django-js-asset==3.1.2
et_xmlfile==2.0.0
kombu==5.5.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
